refactor(github): split client.go by responsibility into 4 new files
Mechanical same-package split of the 1057-line client.go into focused files:
- batch_states.go: repoResult type, FetchRepositoryStates, fetchRepositoryStatesBatch, githubWorker
- ratelimit.go: resetRateLimitAggregation, recordRateLimit, snapshotRateLimit, RateLimitSummary, formatResetLocal
- license_enrich.go: enrichProjectLicenseFromGitHub (pairs with client_license_enrich_test.go)
- commit_stats.go: processCommitHistory, getLatestHumanCommit, getLatestCommit

client.go retains Client/NewClient, FetchBasicRepositoryInfo, FetchDetailedRepositoryInfo,
FetchRepoLanguages, executeGraphQLQuery, graphqlEndpoint, normalizeRepoURL, and repo-info
helpers (MaxTopics, collectTopics, primaryLanguageName, forkSourceFromRepoInfo), reordered
with exported functions preceding unexported ones per project convention.

No signature changes; full test suite passes unchanged.
